Add new generic consumer PathSelectionConsumer
Gzip is to be supported in addition to Zip. A mechanism
for allowing control to pass down the correct path is
necessary to apply differentiated processing depending
upon whether the archive is Zip or Gzip.

(1) We must provide a mechanism to support conditional
processing (choice-of, switch, or content-based-router).
PathSelectionConsumer is that mechanism.
(2) Type ConsumerSelectionAttributes is the configuration
for such a mechanism.
(3) Updated comments.